// src/app/model/member.ts
export class Member {
    id: number;
    picture: String = "https://img.icons8.com/plasticine/2x/user.png";
    name: String = "Jumbo";
    email: String = "jumbotron@gmail.com";
    gender: String = "male";
    phone: number = 9999999999;
    role: String = "UI";
    location: String = "Brazil";
}